Refactored redirections and hooks
