+++ .gitignore
@@ -13,6 +13,7 @@
 !*.m
 !*.cc
 !*.h
+!*.hxx
 !*.rst
 !*.po
 !*.in
